server: Move code in common to server, since it was server specific
